fix last item order
